refactor sstp_protocol -> sstp
