+++ sstp/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "sstp"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ascii = "1.1.0"
+encoding_rs = "0.8.32"
+thiserror = "1.0.40"
+read_macro = { path = "../utils/read_macro" }
+
+[dev-dependencies]
+anyhow = "1.0.71"
+rstest = "0.17.0"
